net: raise MAX_PLAYERS 8 -> 16 (proto 75) - default cap stays 8, 16 opt-in
One binary, structurally widened; runtime cap unchanged (--maxclients 16
opts in, NET_MAX_CLIENTS = 17 with the spectator-host +1).

- constants.h: MAX_PLAYERS 16 (port; N64 keeps 4) + MPCHRSLOTS_PLAYERS/
  BOTS_MASK macros replacing the width literals. MAX_MPCHRS becomes 24,
  per-player arrays auto-grow.
- types.h: g_MpSetup.chrslots + storedbotbits u16 -> u32 (16 player bits
  + 8 bot bits; the options-u64 precedent - unconditional). NOT persisted
  in mpsetups.bin (rebuilt at match start), so no wad version bump.
- Wire: chrslots u16 -> u32 in CLC_ADMIN_SETUP + SVC_STAGE_START;
  NET_PROTOCOL_VER 74 -> 75.
- Mask literal translation, semantics-preserving (current 8-player
  behavior scaled, derived per-site): menutick/challenge/mplayer bot vs
  player splits, pdmain storedbotbits capture, preprocess N64->PC
  chrslots upconvert shift (now MAX_PLAYERS-4).
- g_MpRoomVisibility u16 -> u32 (16 onscreen + 16 standby bits) via the
  MPROOMVIS_* macros from the earlier widening - 3-line change.
- The obj->hidden 4-bit attacker packing holds playernums 0-15 exactly
  (16 values) - sufficient, unchanged; the autogun owner already ships
  as a full client-id byte.

Verified: 9-min Linux soak on the widened binary (8-bot churn, default
cap) OVERALL PASS. Honest limits: 16 simultaneous humans untested (needs
16 real clients); scoreboard/menu layouts may clip beyond 8+8 rows
(cosmetic, follow-up); challenges keep their 4-player semantics.

+++ src/include/types.h
@@ -4187,7 +4187,11 @@ struct mpsetup {
 	 * Bits 0xff00 are for sims
 	 * Externally loaded MP setups get automatically converted to this format.
 	 */
-	/*0x800acb9e*/ u16 chrslots;
+	// Port: widened u16 -> u32 for 16 players + 8 bots (the options-u64
+	// precedent: unconditional on every build; N64 offset comments below are
+	// stale). Not persisted in mpsetups.bin (rebuilt at match start); on the
+	// wire since proto 75 as u32.
+	/*0x800acb9e*/ u32 chrslots;
 	/*0x800acba0*/ u8 weapons[NUM_MPWEAPONSLOTS];
 	/*0x800acba6*/ u8 paused;
 	/*0x800acba8*/ struct fileguid fileguid;
@@ -4205,7 +4209,7 @@ struct mpsetup {
 	u8 racepitytime; // Race: seconds after the first finisher before the match force-ends, 0-120; 0 = end with the winner (port-only)
 	// Used to restore the non-player bits of chrslots upon entering Combat
 	// Simulator, after playing Co-Op/Counter-Op with a human sim.
-	u16 storedbotbits;
+	u32 storedbotbits;
 #endif
 };
